Support the new logger interface
The changes are backwards compatible, but will be blocking for rebar3
since logger uses maps and that means R16 is no longer supported.

Logger's default handler is muted through a filter, and a new handler is
added there with a different name. The same event handler that was used
to add data to error_logger is used, and expanded for logger support as
well.

This seems to make everything work as expected.