Import logging formatter from OTP-21
Earlier versions of OTP-21 exported a version of a formatting function
that was available to everyone. However, following a minor release, this
function was removed (it was never documented, so the risk was kind of
ours).

This in turn meant that the handler for logs in failonly would die after
having to print logs -- the first failing test would work, and the ones
after would stop displaying entries in the shell.

This patch fixes it by wholesale importing the functions that previously
were public within the module, keeping things compatible.